+++ lib/features/auth/sign_up_button.dart
