+++ lib/features/auth/sign_up_success_screen.dart
