+++ lib/features/setting/alarm/notification_setting_container.dart
@@ -2,8 +2,8 @@ import 'package:app_settings/app_settings.dart';
 import 'package:firebase_messaging/firebase_messaging.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:taesung_app/features/setting/features/alarm/widgets/notification_setting_title.dart';
-import 'package:taesung_app/features/setting/features/alarm/widgets/notification_switch.dart';
+import 'package:taesung_app/features/setting/alarm/notification_setting_title.dart';
+import 'package:taesung_app/features/setting/alarm/notification_switch.dart';
 import 'package:taesung_app/models/device_info_model.dart';
 import 'package:taesung_app/models/notification_setting.dart';
 import 'package:taesung_app/providers/notification_setting_provider.dart';
