+++ lib/features/setting/collect/auto_collect_item.dart
