+++ lib/pages/device_alram_setting_page.dart
@@ -1,15 +1,11 @@
 import 'package:flutter/material.dart';
-import 'package:taesung_app/features/setting/features/alarm/widgets/device_alarm_setting_layout.dart';
+import 'package:taesung_app/layouts/device_alarm_setting_layout.dart';
 
 class DeviceAlarmSettingPage extends StatelessWidget {
   const DeviceAlarmSettingPage({super.key});
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-        appBar: AppBar(
-          title: const Text('알람 설정'),
-        ),
-        body: const DeviceAlarmSettingLayout());
+    return const DeviceAlarmSettingLayout();
   }
 }
