+++ lib/pages/home_map_page.dart
@@ -1,13 +1,11 @@
 import 'package:flutter/material.dart';
-import 'package:taesung_app/features/map/map_screen.dart';
+import 'package:taesung_app/layouts/home_map_layout.dart';
 
 class HomeMapPage extends StatelessWidget {
   const HomeMapPage({super.key});
 
   @override
   Widget build(BuildContext context) {
-    return const SafeArea(
-      child: MapScreen(),
-    );
+    return const HomeMapLayout();
   }
 }
