+++ lib/pages/home_setting_page.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
-import 'package:taesung_app/features/setting/widgets/setting_layout.dart';
+import 'package:taesung_app/layouts/home_setting_layout.dart';
 
 class HomeSettingPage extends StatelessWidget {
   const HomeSettingPage({super.key});
 
   @override
   Widget build(BuildContext context) {
-    return const SettingLayout();
+    return const HomeSettingLayout();
   }
 }
