+++ lib/pages/root_page.dart
@@ -1,38 +1,11 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:taesung_app/features/auth/sign_in_page.dart';
-import 'package:taesung_app/pages/home_page.dart';
-import 'package:taesung_app/providers/secure_storage_provider.dart';
+import 'package:taesung_app/layouts/root_layout.dart';
 
-class RootPage extends ConsumerWidget {
+class RootPage extends StatelessWidget {
   const RootPage({super.key});
 
   @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final tokenState = ref.watch(tokenProvider);
-    return tokenState.when(
-      data: (token) {
-        if (token == null) {
-          return const SignInPage();
-        } else {
-          return const HomePage();
-        }
-      },
-      error: (err, st) => Scaffold(
-        body: Column(
-          children: [
-            Text(
-              'Error: $err',
-              style: const TextStyle(color: Colors.red),
-            ),
-          ],
-        ),
-      ),
-      loading: () => const Scaffold(
-        body: Center(
-          child: CircularProgressIndicator(),
-        ),
-      ),
-    );
+  Widget build(BuildContext context) {
+    return const RootLayout();
   }
 }
